According to the current source file, I don't think this issue has been fixed yet even though it's been closed: https://github.com/openrasta/openrasta-stable/issues/19
The problem is with the original 'HttpContext.Current.Request.Path == "/"' statement. The original IF statement is true even when the path is /. This takes away processing from IIS's default document loader; OpenRasta will continue processing if the path is "/".

This is NOT the desired behavior when root default documents are set.

My proposed change simplifies the original hard-to-read IF statement with a more decent check of individual IF short-circuit cancel processing statements. AND correctly exists if the request path is "/" (root).