Add a test around DI testing of Per Web Request
A type registered as transient gets an instance which is created with another instance and is registered as perwebrequest

This is fine in the internal OpenRasta DI
In Castle Windsor DI this fails as it thinks it is waiting for the following dependencies as the was not registered. This is because the registered class depends on something that in not registered, even though there is a valid instance of it.