Moving contexts to OpenRasta.Testing
As a plugin developer, I frequently need to set up parts of OpenRasta
so that I can write unit tests for my extensions.

This commit moves some of the more general contexts to OpenRasta.Testing
with the aim of releasing them as a nuget package to encode best practice
for testing OR extensions.