fix for performance issue: consider handlers as available that can only be resolved dynamically
WindsorDependencyResolver.AvailableHandlers checks among other things
for the CurrentState of a handler to determine its availability.
This is not the best approach, because some handlers can only be
resolved dynamically, e.g. when using DynamicParameters. Those handlers
have the CurrentState "WaitingDependency".
In my case, this causes components registered via OpenRasta to be
registered several times. And since this component is an OpenRasta
Handler, it gets registered with every web-request. This bloats up the
registry and over time slows down the whole application (lookup in a
registry takes longer and longer as it increases).
Checking for the CurrentState is not recommended to be used to check for
available handlers, as also explained here:
http://stackoverflow.com/questions/2732215/how-can-i-validate-the-configuration-of-windsor-castle#tab-top.
I also wrote a test to validate that even if a component might not have
a valid CurrentState, it is still resolveable