Write headers just before the body of a response is written
Instead of deciding when headers are wrote, we detect the first write
to the response stream and send the headers at that point. This would
be followed by an optimisation that would delegate such writing to the
host provided notification of headers being sent, and remove the wrapping
stream, which we'll do later.
Also removes the null check on content length when reading entities, to
support TE.
This was one of the longest standing issues in the pipeline, it closes #101 and the resulting bug #135