Discovered a bug with KeepAlives
I am hosting in IIS using owin.host.systemweb
I have a single client repeatedly GETing and POSTing some resources
I get the error "The server committed a protocol violation. Section=ResponseStatusLine" eventually

Here is a small test app to demonstrate it.  A WebApp which just hosts a single resource, and a Console app which continually GETs and POSTs to the the WebApp.  After running this, an error is shown very quickly